chore: update and adapt feature branch
