+++ packages/core/content-releases/server/src/controllers/release-action.ts
@@ -1,6 +1,6 @@
 import type Koa from 'koa';
 import { validateReleaseActionCreateSchema } from './validation/release-action';
-import type { CreateReleaseAction } from '../../../shared/contracts/release-action';
+import type { CreateReleaseAction } from '../../../shared/contracts/release-actions';
 import { getService } from '../utils';
 
 const releaseActionController = {
