+++ packages/core/content-releases/server/src/services/__tests__/release-validation.test.ts
@@ -1,4 +1,4 @@
-import type { CreateReleaseAction } from '../../../../shared/contracts/release-action';
+import type { CreateReleaseAction } from '../../../../shared/contracts/release-actions';
 import createReleaseValidationService from '../validation';
 
 const baseStrapiMock = {
